# perf(parser): remove backtracking